Create 2025-05-20-My-First-Post.md
Created the file for the first blog post. Mainly tinkered with the YAML format, the usual stuff for the intro, incl. the title and the date. Noted the special requirement for naming the file properly so that it works with Jekyll, and put in some random Lorem Ipsum text for variability.